#752-snapshot-modal - fix snapshot modal window bugs
