+++ src/app/vm/vm-sidebar/storage-detail/volumes/snapshot/snapshot-modal.component.ts
@@ -2,9 +2,9 @@ import {
   Component,
   EventEmitter,
   Input,
+  OnChanges,
   Output
 } from '@angular/core';
-
 import {
   TableDatabase,
   TableDataSource
@@ -21,16 +21,25 @@ import { Snapshot } from '../../../../../shared/models/snapshot.model';
   templateUrl: 'snapshot-modal.component.html',
   styleUrls: ['snapshot-modal.component.scss']
 })
-export class SnapshotModalComponent {
+export class SnapshotModalComponent implements OnChanges {
   public displayedColumns = ['name', 'date', 'actions'];
-  @Input() public dataBase: TableDatabase;
-  @Input() public dataSource: TableDataSource | null;
+  public dataBase: TableDatabase;
+  public dataSource: TableDataSource | null;
   @Input() public volume: Volume;
   @Output() public onSnapshotDelete = new EventEmitter<Snapshot>();
 
   constructor(
     public snapshotActionsService: SnapshotActionsService,
   ) { }
+
+  public ngOnChanges() {
+    this.update();
+  }
+
+  public update() {
+    this.dataBase = new TableDatabase(this.volume.snapshots);
+    this.dataSource = new TableDataSource(this.dataBase);
+  }
 
   public onAction(action: SnapshotAction, snapshot: Snapshot) {
     action.activate(snapshot).subscribe(
