+++ src/app/vm/vm-sidebar/storage-detail/volumes/snapshot/snapshots.component.ts
@@ -33,7 +33,7 @@ export class SnapshotsComponent {
 
   public showSnapshots(): void {
     this.dialog.open(SnapshotModalContainerComponent, {
-      data: { volume: this.volume },
+      data: { volumeId: this.volume.id },
       width: '700px'
     }).afterClosed();
   }
